// src/types.ts
export type SchoolUserRole = 'student' | 'teacher' | 'head_of_department' | 'canteen_seller' | 'deputy_headteacher' | 'parent' | 'old_student' | 'carrier' | 'admin' | 'admission_agent';

export interface User {
  name: string;
  studentId: string;
  class?: string;
  stream?: string;
  schoolId?: string;
  role: 'student' | 'superadmin' | SchoolUserRole;
  password?: string;
  avatarUrl?: string;
  contactNumber?: string;
  email?: string;
  bio?: string;
  mustChangePassword?: boolean;
  dateOfBirth?: string;
  address?: string;
  gender?: 'Male' | 'Female';
  shopId?: string; // For canteen_seller
  unebPassSlip?: UnebPassSlip; // To store verified results
  internalExams?: InternalExamResult[];
  accountStatus?: 'temporary' | 'active' | 'disabled';
  pendingTransferAcceptance?: boolean;
  teachingAssignments?: { [className: string]: string[] };
}

export interface AdminUser {
  id: string;
  name: string;
  email: string;
  role: 'headteacher' | 'uneb_admin' | 'nche_admin';
  assignedSchoolIds: string[];
  password: string;
  avatarUrl?: string;
  contactNumber?: string;
  bio?: string;
  lastLogin?: number;
  address?: string;
}


// --- School & Module Types ---
export interface School {
  id: string;
  name: string;
  address: string;
  modules: { 
    moduleId: string; 
    status: 'assigned' | 'active' | 'published';
    allowedRoles?: SchoolUserRole[];
  }[];
  isHomePagePublished?: boolean;
}
// ...
export interface UnebPassSlip {
  indexNo: string;
  name: string;
  year: string;
  level: 'P.L.E' | 'U.C.E' | 'U.A.C.E';
  subjects: { name: string; grade: string }[];
  dateOfBirth?: string;
  schoolName?: string;
  schoolAddress?: string;
  entryCode?: string;
  aggregate?: string;
  result?: string;
}
// ...
export interface SubjectPerformance {
  name: string;
  score: number; // Percentage score
  grade: string;
  remarks?: string;
}

export interface InternalExamResult {
  term: string; // e.g., "Term 1, 2024"
  subjects: SubjectPerformance[];
  average: number;
  classPosition: string; // e.g., "3rd out of 45"
}